% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mum_GUI.R
\name{mum_GUI}
\alias{mum_GUI}
\title{Online Graphical User Interface to compute MUMARINEX}
\usage{
mum_GUI()
}
\value{
Opens the web browser to the online Shiny application.
}
\description{
Launch the online Graphical User Interface to compute the MUMARINEX index and its sub-indices.
}
\details{
mum_GUI
}
\examples{
mum_GUI()
}
